+++ src/services/owncloudservice.cpp
@@ -969,6 +969,11 @@ bool OwnCloudService::updateICSDataOfCalendarItem(CalendarItem *calItem) {
     // 5 sec timeout for the request
     timer.start(5000);
 
+#if (QT_VERSION >= QT_VERSION_CHECK(5, 6, 0))
+    r.setAttribute(QNetworkRequest::FollowRedirectsAttribute,
+                                true);
+#endif
+
     QNetworkReply *reply = manager->get(r);
     ignoreSslErrorsIfAllowed(reply);